Moved DealDataResult to storage
