+++ packages/core/src/upload-manager.ts
@@ -24,14 +24,6 @@ export interface UploadResult {
   cid?: string;
 }
 
-export interface DealDataResult {
-  pieceSize: number;
-  size: number;
-  pieceCid: string;
-  dataCid: string;
-  carLink: string;
-}
-
 class UploadManager {
   private readonly spheronApiUrl: string = "https://api-v2.spheron.network";
