+++ packages/storage/src/index.ts
@@ -13,15 +13,14 @@ import {
   TokenScope,
   UploadManager,
   UploadResult,
-  PinStatus,
-  DealDataResult,
+  PinStatus,  
 } from "@spheron/core";
 import { createPayloads, processCarFile } from "./fs-payload-creator";
 import { ipfs } from "./ipfs.utils";
 import { UsageWithLimits } from "./bucket-manager/interfaces";
 import fs from 'fs';
 
-import { DecryptFromIpfsProps, EncryptToIpfsProps } from "./interface";
+import { DecryptFromIpfsProps, EncryptToIpfsProps, DealDataResult } from "./interface";
 import {
   uint8arrayFromString,
   encryptData,
